Backend/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
web3==5.31.3
python-dotenv==1.0.0
Werkzeug==2.2.3
pytest==7.4.0
pytest-mock==3.11.1